Further simplified and reorganized the new code.
