+++ src/application/views/icon.ico
