+++ src/application/views/registration.py
